Update README.md
Mad Libs Game

Create your own Mad Libs by writing a program that plays a word game with the user. It should have a similar format to the example below.

Important Note: Your Mad Libs should be unique and NOT be the same as the example below.

Here is a description of an example program. First, the user is asked to enter the following:

His or her first name
His or her age
The name of a city
The name of a planet
A profession
A type of animal
A pet’s name
After the user has entered these items, the program displays the following story, inserting the user’s input into the appropriate locations:

There once was a person named NAME who lived in CITY. At the age of AGE, NAME moved to the planet PLANET. There, NAME worked as a PROFESSION. Over time, NAME adopted a(n) ANIMAL named PETNAME and they both traveled the universe.